TypeSystemContext: introduce separate typeDepthForApproximation
Before this commit, we used TypeSystemContext.typeDepth for various
purposes. But it looks like that in different situations the rules
should be slightly different, for example, it's important to know
depth of a captured type during approximation, but it's not important,
and even dangerous (can cause infinite recursion) during constraint
injection. So here we keep the general-purpose typeDepth() and
introduce the separate typeDepthForApproximation() that fallbacks by
default to the general one.

Related to KT-60034, KT-28598