[minor] Fix the configuration order of UpdateTestDataHandler
Before this commit, it was possible for some after-analysis checkers to be executed after `UpdateTestDataHandler` due to the existence of various configuration callbacks.

For example, `FirFailingTestSuppressor` is added to a number of test configurations per regex-matching test data directory, which is performed via callbacks that are executed way after calls to the primary `configure` and `TestConfigurationBuilder.configuration` functions of `AbstractKotlinCompilerTest`. This has effectively led to `UpdateTestDataHandler` ignoring the presence of `.fir.fail` files.

This commit moves the addition of `UpdateTestDataHandler` to test configurations to the very last moment before a `TestConfiguration` is actually created to prevent situations like the one described above from happening.